+++ scout/adapter/mongo/query.py
@@ -298,7 +298,7 @@ class QueryHandler(object):
         # there is only 'clinsig' criterion among the primary terms right now
         primary_terms = False
 
-        # gnomad_frequency, local_obs, local_obs_freq, clingen_ngi, swegen, spidex_human, cadd_score, genetic_models, mvl_tag, clinvar_tag, cosmic_tag
+        # gnomad_frequency, local_obs, local_obs_freq, clingen_ngi, swegen, swegen_freq, spidex_human, cadd_score, genetic_models, mvl_tag, clinvar_tag, cosmic_tag
         # functional_annotations, region_annotations, size, svtype, decipher, depth, alt_count, control_frequency, tumor_frequency
         secondary_terms = False
 
@@ -564,19 +564,14 @@ class QueryHandler(object):
 
             if criterion == "gnomad_frequency":
                 gnomad = query.get("gnomad_frequency")
-                if gnomad == "-1":
-                    # -1 means to exclude all variants that exists in gnomad
-                    mongo_query["gnomad_frequency"] = {"$exists": False}
-                else:
-                    # Replace comma with dot
-                    mongo_secondary_query.append(
-                        {
-                            "$or": [
-                                {"gnomad_frequency": {"$lt": float(gnomad)}},
-                                {"gnomad_frequency": {"$exists": False}},
-                            ]
-                        }
-                    )
+                mongo_secondary_query.append(
+                    {
+                        "$or": [
+                            {"gnomad_frequency": {"$lt": float(gnomad)}},
+                            {"gnomad_frequency": {"$exists": False}},
+                        ]
+                    }
+                )
 
             if criterion == "local_obs":
                 local_obs = query.get("local_obs")
@@ -596,6 +591,17 @@ class QueryHandler(object):
                         "$or": [
                             {"local_obs_old_freq": None},
                             {"local_obs_old_freq": {"$lt": local_obs_freq}},
+                        ]
+                    }
+                )
+
+            if criterion == "swegen_freq":
+                swegen = query.get("swegen_freq")
+                mongo_secondary_query.append(
+                    {
+                        "$or": [
+                            {"swegen_mei_max": {"$lt": float(swegen)}},
+                            {"swegen_mei_max": {"$exists": False}},
                         ]
                     }
                 )